% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonize.R
\name{codonize}
\alias{codonize}
\title{codonize}
\usage{
codonize(sequence, start = 1)
}
\arguments{
\item{sequence}{DNA or RNA sequence}

\item{start}{codon start position}
}
\value{
codonize returns the codons of a DNA or RNA sequence
}
\description{
codonize
}
\examples{
codonize(sequence = "ATGACATCGAUGTTA", start = 1)
}
