% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_2.R
\name{function2}
\alias{function2}
\title{Transcription}
\usage{
function2(x2)
}
\arguments{
\item{x2}{this function replace T by U}
}
\value{
It returns the same sequence but changing T by U
}
\description{
Transcription
}
