% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDNA.R
\name{makeDNA}
\alias{makeDNA}
\title{Make a DNA sequence}
\usage{
makeDNA(generateseq)
}
\arguments{
\item{generateseq}{this function generates a sampled line of nucleotides, in a DNA_string format.}
}
\value{
the function returns a character string of DNA bases.
}
\description{
Make a DNA sequence
}
