% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aminoacids_abundance.R
\name{plot_aminoacids_abundance}
\alias{plot_aminoacids_abundance}
\title{make bar plot of aminoacids abundance}
\usage{
plot_aminoacids_abundance(aminoacid_sequence)
}
\arguments{
\item{aminoacid_sequence}{sequence of aminoacids}
}
\value{
bar plot of the aminoacids abundance
}
\description{
make bar plot of aminoacids abundance
}
