% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_aminoacid_translation.R
\name{RNA_aminoacid_translation}
\alias{RNA_aminoacid_translation}
\title{RNA_aminoacid_translation}
\usage{
RNA_aminoacid_translation(codons)
}
\arguments{
\item{codons}{selected RNA sequence}
}
\value{
string of Aminoacids out of the selected RNA
}
\description{
RNA_aminoacid_translation
}
